restructure root docs app
