+++ scopes/react/ui/docs-app/docs-app.tsx
@@ -1,27 +1,55 @@
-// import 'reset-css'; // do not include resets, we want compositions with native behavior
-import React, { PropsWithChildren } from 'react';
-import { ThemeContext } from '@teambit/documenter.theme.theme-context';
-import { IconFont } from '@teambit/design.theme.icons-font';
 
-export type DocsAppProps = {
-  Provider?: React.ComponentType;
-  children: React.ReactChild
+import React from 'react';
+import { isFunction } from 'lodash';
+import { Properties } from '@teambit/react.ui.properties';
+import { RenderingContext } from '@teambit/preview';
+import { CompositionsSummary } from '@teambit/react.ui.compositions-summary';
+import { DocsContent } from '@teambit/react.ui.docs-content';
+import { ApplyProviders } from '@teambit/react.ui.apply-providers';
+
+import { DocsTheme } from './docs-theme';
+import { ExamplesOverview } from './examples-overview';
+import type { DocsFile } from './examples-overview';
+import styles from './docs-app.module.scss';
+
+export type ReactDocsAppParams = {
+  componentId: string,
+  docs: DocsFile | undefined,
+  compositions: any,
+  context: RenderingContext
+};
+
+const defaultDocs = {
+  examples: [],
+  labels: [],
+  abstract: '',
 };
 
 export function DocsApp({
-  Provider = Noop,
-  children
-}: DocsAppProps) {
-  return (
-    <Provider>
-      <ThemeContext>
-        <IconFont query="q76y7n" />
-        {children}
-      </ThemeContext>
-    </Provider>
-  );
-}
+    componentId,
+    docs = defaultDocs,
+    compositions,
+    context
+  }: ReactDocsAppParams
+) {
 
-function Noop({ children }: PropsWithChildren<{}>) {
-  return <>{children}</>;
+  // Next 2 lines are to support legacy code (ExamplesOverview)
+  const { examples = [] } = docs;
+  const Content: any = isFunction(docs.default) ? docs.default : () => null;
+
+  return (
+    <DocsTheme>
+      <ApplyProviders renderingContext={context}>
+        <DocsContent docs={docs} className={styles.mdx}/>
+
+        <CompositionsSummary
+          compositions={compositions}
+          className={styles.compositionSection}
+          compositionCardClass={styles.compositionCard}
+        />
+        <Properties componentId={componentId} />
+        <ExamplesOverview examples={Content.examples || examples} />
+      </ApplyProviders>
+    </DocsTheme>
+  );
 }
