+++ scopes/react/ui/docs-app/docs-theme.tsx
@@ -1,0 +1,19 @@
+// import 'reset-css'; // do not include resets, we want compositions with native behavior
+import React from 'react';
+import { ThemeContext } from '@teambit/documenter.theme.theme-context';
+import { IconFont } from '@teambit/design.theme.icons-font';
+
+export type DocsAppProps = {
+  children: React.ReactChild
+};
+
+export function DocsTheme({
+  children
+}: DocsAppProps) {
+  return (
+    <ThemeContext>
+      <IconFont query="q76y7n" />
+      {children}
+    </ThemeContext>
+  );
+}
